Set authentication object when resuming exercise because of new commit.
This sets an authentication object when an exercise is resumed automatically because a new commit was received. The logic for setting the authentication object is now handled in a separate method, as it is required more than once.
